Basic species creation and population working
The factories for species and individual creature creation are working
now on a basic level, and the world can now easily populate itself with
creatures with just a few method calls.  The factories for creating
species, the lists that contain the individual creatures (instances) of
the species, and the food chain tree that shows the predator/prey
relation between species are now all tied together with an integer id
system.