Implemented form of breeding
Creatures can now breed, and a strategy pattern has been set up so the
breeding behavior can be swapped out at runtime if desired.  Currently
the breeding pattern is that each creature has a chance to make one
baby according to it’s species’ birthrate.  This is a very simple
approximation, as creatures don’t currently have to couple up to breed,
so I’m temporarily dividing each birthrate in half to account for this.
 Also the world object is being passed around a lot, so I should start
finding out a way to cut down on that because that many classes
shouldn’t have access to all of the world methods…